+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.2.3 - Unreleased
 
+### Fixed
+* Issue with slurm 21.x
+
 
 ## Version 0.2.2 - Released 2025-02-19